Utils move to src from component
