+++ client/src/components/Auth/SignIn/SignIn.js
@@ -4,7 +4,7 @@ import {toast} from 'react-toastify';
 import styles from './SignIn.module.css';
 import fb from './fb.png';
 import gog from './gog.png';
-import * as validator from '../../utils/validator';
+import * as validator from '../../../utils/validator';
 
 function SignIn() {
